HyperShift: Move hypershift.go to platform package
This commit doesn't introduce functional changes.
Move hypershift.go to platform package to avoid circural dependencies in following commits.
