+++ pkg/controller/statusmanager/status_manager.go
@@ -3,7 +3,6 @@ package statusmanager
 import (
 	"context"
 	"fmt"
-	"k8s.io/apimachinery/pkg/labels"
 	"log"
 	"os"
 	"reflect"
@@ -17,7 +16,7 @@ import (
 	"github.com/openshift/cluster-network-operator/pkg/apply"
 	cnoclient "github.com/openshift/cluster-network-operator/pkg/client"
 	"github.com/openshift/cluster-network-operator/pkg/names"
-	"github.com/openshift/cluster-network-operator/pkg/network"
+	"github.com/openshift/cluster-network-operator/pkg/platform"
 	hyperv1 "github.com/openshift/hypershift/api/v1beta1"
 	cohelpers "github.com/openshift/library-go/pkg/config/clusteroperator/v1helpers"
 	operstatus "github.com/openshift/library-go/pkg/operator/status"
@@ -28,6 +27,7 @@ import (
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	uns "k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/tools/cache"
@@ -82,7 +82,7 @@ type StatusManager struct {
 
 	client           cnoclient.Client
 	name             string
-	hyperShiftConfig *network.HyperShiftConfig
+	hyperShiftConfig *platform.HyperShiftConfig
 
 	failing         [maxStatusLevel]*operv1.OperatorCondition
 	installComplete bool
@@ -106,7 +106,7 @@ func New(client cnoclient.Client, name, cluster string) *StatusManager {
 	status := &StatusManager{
 		client:           client,
 		name:             name,
-		hyperShiftConfig: network.NewHyperShiftConfig(),
+		hyperShiftConfig: platform.NewHyperShiftConfig(),
 
 		dsInformers:  map[string]cache.SharedIndexInformer{},
 		dsListers:    map[string]DaemonSetLister{},
@@ -134,10 +134,10 @@ func (status *StatusManager) setClusterOperAnnotation(obj *configv1.ClusterOpera
 }
 
 // getClusterOperAnnotation gets an annotation from the clusterOperator network object
-func (status *StatusManager) getClusterOperAnnotation(obj *configv1.ClusterOperator) ([]network.RelatedObject, error) {
+func (status *StatusManager) getClusterOperAnnotation(obj *configv1.ClusterOperator) ([]platform.RelatedObject, error) {
 	new := obj.DeepCopy()
 	anno := new.GetAnnotations()
-	objs := []network.RelatedObject{}
+	objs := []platform.RelatedObject{}
 
 	value, set := anno[names.RelatedClusterObjectsAnnotation]
 	if !set || value == "" {
@@ -153,7 +153,7 @@ func (status *StatusManager) getClusterOperAnnotation(obj *configv1.ClusterOpera
 			return objs, fmt.Errorf("'%s' annotation is invalid, expected: ClusterName/Group/Resource/Namespace/Name, got: %s",
 				names.RelatedClusterObjectsAnnotation, res)
 		}
-		objs = append(objs, network.RelatedObject{
+		objs = append(objs, platform.RelatedObject{
 			ClusterName: parts[0],
 			ObjectReference: configv1.ObjectReference{
 				Group:     parts[1],
@@ -300,7 +300,7 @@ func (status *StatusManager) writeHypershiftStatus(operStatus *operv1.NetworkSta
 				}
 
 				newCondition := metav1.Condition{
-					Type:    network.HyperShiftConditionTypePrefix + cond.Type,
+					Type:    platform.HyperShiftConditionTypePrefix + cond.Type,
 					Status:  metav1.ConditionStatus(cond.Status),
 					Reason:  reason,
 					Message: cond.Message,
@@ -584,7 +584,7 @@ func (status *StatusManager) SetRelatedObjects(relatedObjects []configv1.ObjectR
 	status.relatedObjects = relatedObjects
 }
 
-func (status *StatusManager) SetRelatedClusterObjects(relatedObjects []network.RelatedObject) {
+func (status *StatusManager) SetRelatedClusterObjects(relatedObjects []platform.RelatedObject) {
 	status.Lock()
 	defer status.Unlock()
 	status.hyperShiftConfig.RelatedObjects = relatedObjects
